improvements to delete application page

+++ Code/applicantHome.php
@@ -82,8 +82,8 @@ mysqli_close($db_connection);
 $body = <<<BODY
     <h1>TA Application</h1>
     <a class="btn btn-primary" href="apply1.php">Apply/Edit Application</a>
-    <a class="btn btn-primary" href="deleteApplication.php">Remove Application</a>
-    <a class="btn btn-primary" href="displayApplication.php/?id=$directoryID">Review Application</a>
+    <a class="btn btn-primary" href="deleteApplication.php" onclick = "return confirm('Are you sure you want to remove your application?');">Remove Application</a>
+    <a class="btn btn-primary" href="applicantViewApplication.php">Review Application</a>
 
 
 BODY;
